Move iterator code into tenant_shard_iterator.rs
